+++ src/DagAir_IngestionNode/DagAir.IngestionNode.Data/Measurements/InfluxRoomMeasurement.cs
@@ -8,9 +8,9 @@ namespace DagAir.IngestionNode.Data.Measurements
     public class InfluxRoomMeasurement
     {
         [Column("sensor_id", IsTag = true)] public string SensorId { get; set; }
-        [Column("temperature")] public double? Temperature { get; set; }
-        [Column("humidity")] public double? Humidity { get; set; }
-        [Column("illuminance")] public double? Illuminance { get; set; }
+        [Column("temperature")] public decimal? Temperature { get; set; }
+        [Column("humidity")] public decimal? Humidity { get; set; }
+        [Column("illuminance")] public int? Illuminance { get; set; }
         [Column(IsTimestamp = true)] public DateTime Time { get; set; }
     }
 }
