+++ src/DagAir_Policies/DagAir.Policies.Contracts/DTOs/ExpectedRoomConditionsDto.cs
@@ -5,12 +5,12 @@ namespace DagAir.Policies.Contracts.DTOs
     public class ExpectedRoomConditionsDto
     {
         public long Id { get; set; }
-        public float Temperature { get; set; }
-        public float Illuminance { get; set; } //light intensity
-        public float Humidity { get; set; }
-        public float TemperatureMargin { get; set; }
-        public float IlluminanceMargin { get; set; }
-        public float HumidityMargin { get; set; }
+        public decimal Temperature { get; set; }
+        public int Illuminance { get; set; } //light intensity
+        public decimal Humidity { get; set; }
+        public decimal TemperatureMargin { get; set; }
+        public decimal IlluminanceMargin { get; set; }
+        public decimal HumidityMargin { get; set; }
         public long? RoomPolicyId { get; set; }
         public List<RoomPolicyDto> RoomPolicies { get; set; }
     }
