+++ src/DagAir_Policies/DagAir.Policies.Tests/Policies/PoliciesEndpointPathsTests.cs
@@ -91,10 +91,10 @@ namespace DagAir.Policies.Tests.Policies
                 Id = 1,
                 Temperature = 20,
                 Illuminance = 100,
-                Humidity = 0.4f,
+                Humidity = (decimal) 0.4,
                 TemperatureMargin = 2,
                 IlluminanceMargin = 20,
-                HumidityMargin = 0.1f
+                HumidityMargin = (decimal) 0.1
             };
 
             await AppContext.ExpectedRoomConditions.AddAsync(expectedRoomConditions);
